requirements.txt
djangorestframework==3.6.2
django==1.11
braintree==3.39.0
